CONFLUENCE-62: Confluence XML importer fails on attachments with IDs above Integer.MAX_VALUE

+++ confluence-xml/src/main/java/org/xwiki/contrib/confluence/filter/internal/input/ConfluenceInputFilterStream.java
@@ -613,7 +613,7 @@ public class ConfluenceInputFilterStream
             return;
         }
 
-        long attachmentId = attachmentProperties.getInt("id");
+        long attachmentId = attachmentProperties.getLong("id");
 
         String attachmentName = this.confluencePackage.getAttachmentName(attachmentProperties);
